+++ src/interactors/buildTable/sortByStats.js
@@ -1,10 +1,12 @@
 const { STATS } = require('../../config/stats');
+const { isNil } = require('../../utils');
 
 const buildSort = (statConfig) => (a, b) => {
   const { id, sortOrder } = statConfig;
   const { stats: statsA = {} } = a;
   const { stats: statsB = {} } = b;
   const multiplier = sortOrder === 'DESC' ? -1 : 1;
+  if (isNil(statsA[id])) return 1;
   return multiplier * (statsA[id] - statsB[id]);
 };