+++ src/interactors/getReviewStats/__tests__/calculateReviewsStats.test.js
@@ -1,8 +1,9 @@
-const { reviews: input } = require('../../../../tests/mocks');
+const { reviews: input, pullRequests } = require('../../../../tests/mocks');
 const calculateReviewsStats = require('../calculateReviewsStats');
 
 describe('Interactors | getReviewStats | .calculateReviewsStats', () => {
-  const result = calculateReviewsStats(input);
+  const pullsById = pullRequests.reduce((acc, pr) => ({ ...acc, [pr.id]: pr }), {});
+  const result = calculateReviewsStats(input, pullsById);
 
   it('calculates the totalReviews', () => {
     expect(result.totalReviews).toBe(2);
@@ -12,11 +13,23 @@ describe('Interactors | getReviewStats | .calculateReviewsStats', () => {
     expect(result.totalComments).toBe(6);
   });
 
-  it('calculates the commentsPerReview', () => {
-    expect(result.commentsPerReview).toBe(3);
-  });
-
   it('calculates the timeToReview', () => {
     expect(result.timeToReview).toBe(75000);
+  });
+
+  it('calculates the commentsPerReview', () => {
+    expect(result.commentsPerReview).toBe(2);
+  });
+
+  it('calculates the reviewedAdditions', () => {
+    expect(result.reviewedAdditions).toBe(173);
+  });
+
+  it('calculates the reviewedDeletions', () => {
+    expect(result.reviewedDeletions).toBe(87);
+  });
+
+  it('calculates the reviewedLines', () => {
+    expect(result.reviewedLines).toBe(260);
   });
 });