+++ src/interactors/getReviewStats/index.js
@@ -1,8 +1,12 @@
 const calculateReviewsStats = require('./calculateReviewsStats');
 const groupReviews = require('./groupReviews');
 
-module.exports = (pulls) => groupReviews(pulls)
-  .map(({ userId, reviews }) => {
-    const stats = calculateReviewsStats(reviews);
-    return { userId, reviews, stats };
-  });
+module.exports = (pulls) => {
+  const pullsById = pulls.reduce((acc, pull) => ({ ...acc, [pull.id]: pull }), {});
+
+  return groupReviews(pulls)
+    .map(({ userId, reviews }) => {
+      const stats = calculateReviewsStats(reviews, pullsById);
+      return { userId, reviews, stats };
+    });
+};