+++ src/parsers/parsePullRequest.js
@@ -9,13 +9,18 @@ const getFilteredReviews = (data) => get(data, 'node.reviews.nodes', []).filter(
 module.exports = (data = {}) => {
   const author = parseUser(get(data, 'node.author'));
   const publishedAt = new Date(get(data, 'node.publishedAt'));
+  const additions = get(data, 'node.additions');
+  const deletions = get(data, 'node.deletions');
   const handleReviews = (review) => parseReview(review, { publishedAt, authorLogin: author.login });
 
   return {
     author,
+    additions,
+    deletions,
     publishedAt,
     cursor: data.cursor,
     id: get(data, 'node.id'),
+    lines: additions + deletions,
     reviews: getFilteredReviews(data).map(handleReviews),
   };
 };